fix: widen mission entity/group ids to ulong, derive squad slot, add world-unit conversion
Task 64: MissionState/MissionSnapshot/SandataStateHasher were the last
Sandata.Core types still using int for entity and group identifiers,
disagreeing with Hukbo.Core.Simulation.AgentState.EntityId and Sandata's own
SandataCollisionBody/SandataCollisionPair/SandataCollisionMoveRequest. Widen
OperatorState.EntityId, ContactMemoryEntry.EnemyEntityId, GroupPathState.GroupId,
and MissionState/MissionSnapshot.NextEntityId to ulong; leave FactionId,
DoorId, cell indices, and RNG stream/algorithm ids as int since none of them
name an entity or a group.

Remove OperatorState.SquadSlotIndex per design section 4's 2026-08-07
correction: design section 8 derives group id, leader, membership, and slot
index every tick and stores nothing per group, so the field contradicted the
design it was supposed to implement.

Add Sandata.Core.Mathematics.WorldUnits, the missing FixedPoint-to-world-unit
bridge task 32 needed and reported absent: floors via arithmetic right shift
by log2(FixedPoint.Scale), agreeing with NavGrid.WorldToCellCoordinate's own
shift-based floor rule rather than C# division's truncation toward zero.

No golden mission hash exists yet, so no pinned constant moved.

+++ src/Sandata.Core/Simulation/MissionSnapshot.cs
@@ -20,7 +20,7 @@ public sealed record MissionSnapshot(
     long Tick,
     int Phase,
     int Winner,
-    int NextEntityId,
+    ulong NextEntityId,
     long NextEventSequence)
 {
     public ImmutableArray<OperatorState> Operators { get; init; } =
